# app/schemas/chat.py
from datetime import datetime
from typing import Any
from uuid import UUID

from pydantic import BaseModel, ConfigDict, Field

from app.models import MessageRole, TicketStatus
from app.schemas.chatbot import WidgetTheme

# The widget's session id, in one place. It reaches the API three ways — a chat body, a
# ticket body and a bootstrap query string — and a looser rule on any one of them would be a
# looser way in, so all three are built from these.
SESSION_ID_MIN_LENGTH = 8
SESSION_ID_MAX_LENGTH = 128
SESSION_ID_PATTERN = r"^[A-Za-z0-9_\-]+$"


def session_id_field() -> Any:
    return Field(
        min_length=SESSION_ID_MIN_LENGTH,
        max_length=SESSION_ID_MAX_LENGTH,
        pattern=SESSION_ID_PATTERN,
    )


class WidgetChatRequest(BaseModel):
    message: str = Field(min_length=1, max_length=4000)
    # Generated by the widget and stored in the host page's storage. Anonymous by design:
    # chatting with a bot must not require identifying the end user.
    session_id: str = session_id_field()


class SourceRead(BaseModel):
    marker: int
    chunk_id: str
    document_id: str
    similarity: float
    excerpt: str
    metadata: dict[str, Any] = Field(default_factory=dict)


class MessageRead(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: UUID
    conversation_id: UUID
    role: MessageRole
    content: str
    sources_json: list[dict[str, Any]] | None
    latency_ms: int | None
    created_at: datetime
    # Set only on `role='staff'` turns. Dashboard-only: the widget is served `WidgetMessage`,
    # which has no such field, so a visitor never learns which colleague answered them.
    staff_user_id: UUID | None = None


class ConversationRead(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: UUID
    chatbot_id: UUID
    external_session_id: str
    title: str | None
    message_count: int
    created_at: datetime
    updated_at: datetime


class WidgetMessage(BaseModel):
    """One transcript turn as the *visitor* is allowed to see it on their way back in.

    Deliberately narrower than `MessageRead`: no ids, no token counts, no latency, and above
    all no `staff_user_id`. A visitor learns that a human replied, never which one.
    """

    model_config = ConfigDict(from_attributes=True)

    role: MessageRole
    content: str
    created_at: datetime


class WidgetSessionState(BaseModel):
    """What a returning visitor gets back when bootstrap is given their session id.

    There is no outbound mail transport, so this is the only way a staff reply reaches the
    person who asked for it: they open the widget again and it is in the log.

    `ticket_status` is the whole of what leaks about the ticket — not the address they gave,
    not who it was assigned to, not its priority. Those belong to the organisation, and the
    session id is a bearer capability rather than an authenticated identity.
    """

    conversation_id: UUID
    messages: list[WidgetMessage]
    ticket_status: TicketStatus | None = None


class WidgetBootstrap(BaseModel):
    """Everything the widget needs to render before its first message.

    `name` and `greeting` are already resolved against the theme's overrides, so the widget
    renders what it is given rather than deciding which of two sources wins.
    """

    chatbot_id: UUID
    name: str
    greeting: str
    status: str
    theme: WidgetTheme
    # Footer links, empty when the tenant has set none. Separate from `theme` because they
    # are the tenant's legal disclosures rather than their colours, and survive a theme reset.
    privacy_url: str = ""
    terms_url: str = ""
    # Absent on a first visit and whenever no session id was supplied, which is what keeps
    # the greeting-only path exactly as it was.
    session: WidgetSessionState | None = None