# frontend/packages/types/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageRead": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "role": {
            "$ref": "#/components/schemas/MessageRole"
          },
          "sources_json": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources Json"
          },
          "staff_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staff User Id"
          }
        },
        "required": [
          "id",
          "conversation_id",
          "role",
          "content",
          "sources_json",
          "latency_ms",
          "created_at"
        ],
        "title": "MessageRead",
        "type": "object"
      },
      "MessageRole": {
        "enum": [
          "user",
          "assistant",
          "staff"
        ],
        "title": "MessageRole",
        "type": "string"
      },
...
      "Page_TicketRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TicketRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[TicketRead]",
        "type": "object"
      },
...
      "TicketCreate": {
        "description": "What the widget's contact form posts.\n\n`conversation_id` is absent when the visitor asked for a human before saying anything, in\nwhich case the conversation is created from `session_id` the same way a first message\nwould create it.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "escalation_reason": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Reason"
          },
          "message": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "session_id": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9_\\-]+$",
            "title": "Session Id",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/TicketSource",
            "default": "visitor_contact_form"
          },
          "subject": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          }
        },
        "required": [
          "email",
          "session_id"
        ],
        "title": "TicketCreate",
        "type": "object"
      },
      "TicketCreated": {
        "description": "Deliberately thin.\n\nThe widget needs to know the request landed and which conversation it belongs to; it has\nno business knowing who the ticket was routed to or what else is in the queue.",
        "properties": {
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "ticket_id": {
            "format": "uuid",
            "title": "Ticket Id",
            "type": "string"
          }
        },
        "required": [
          "ticket_id",
          "conversation_id",
          "status"
        ],
        "title": "TicketCreated",
        "type": "object"
      },
      "TicketDetail": {
        "description": "A ticket plus the conversation it wraps, which is where the thread actually lives.",
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageRead"
            },
            "title": "Messages",
            "type": "array"
          },
          "ticket": {
            "$ref": "#/components/schemas/TicketRead"
          }
        },
        "required": [
          "ticket",
          "messages"
        ],
        "title": "TicketDetail",
        "type": "object"
      },
      "TicketPriority": {
        "enum": [
          "low",
          "normal",
          "high",
          "urgent"
        ],
        "title": "TicketPriority",
        "type": "string"
      },
      "TicketRead": {
        "properties": {
          "assigned_to": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "chatbot_id": {
            "format": "uuid",
            "title": "Chatbot Id",
            "type": "string"
          },
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "escalation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Reason"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "org_id": {
            "format": "uuid",
            "title": "Org Id",
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/TicketPriority"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "source": {
            "$ref": "#/components/schemas/TicketSource"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "visitor_email": {
            "title": "Visitor Email",
            "type": "string"
          },
          "visitor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visitor Name"
          }
        },
        "required": [
          "id",
          "org_id",
          "chatbot_id",
          "conversation_id",
          "visitor_email",
          "visitor_name",
          "subject",
          "status",
          "priority",
          "source",
          "escalation_reason",
          "assigned_to",
          "resolved_at",
          "created_at",
          "updated_at"
        ],
        "title": "TicketRead",
        "type": "object"
      },
      "TicketReply": {
        "properties": {
          "content": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "TicketReply",
        "type": "object"
      },
      "TicketSource": {
        "description": "How the visitor got here: offered the option after a grounding miss, or they asked.",
        "enum": [
          "ai_escalation",
          "visitor_contact_form"
        ],
        "title": "TicketSource",
        "type": "string"
      },
      "TicketStatus": {
        "description": "`open` = new and unclaimed. `pending` = assigned, or a staff member has replied at\nleast once. `resolved` = done, with `resolved_at` set. `closed` = archived terminal\nstate. Reopening from either terminal state clears `resolved_at`.",
        "enum": [
          "open",
          "pending",
          "resolved",
          "closed"
        ],
        "title": "TicketStatus",
        "type": "string"
      },
      "TicketUpdate": {
        "description": "Every field optional; `assigned_to` distinguishes \"unassign\" from \"leave alone\".\n\nPydantic cannot tell an omitted nullable field from one explicitly set to null, so\nunassignment is its own flag rather than a null that would be indistinguishable from\nsilence.",
        "properties": {
          "assigned_to": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TicketPriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TicketStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "unassign": {
            "default": false,
            "title": "Unassign",
            "type": "boolean"
          }
        },
        "title": "TicketUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Dashboard access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "RAG Chatbot Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tickets": {
      "get": {
        "description": "Readable by any member: the support queue is shared work, not privileged information.",
        "operationId": "list_tickets_api_v1_tickets_get",
        "parameters": [
          {
            "in": "query",
            "name": "chatbot_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chatbot Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TicketStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "assigned_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assigned To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TicketRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Tickets",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/v1/tickets/{ticket_id}": {
      "get": {
        "operationId": "get_ticket_api_v1_tickets__ticket_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Ticket",
        "tags": [
          "tickets"
        ]
      },
      "patch": {
        "operationId": "update_ticket_api_v1_tickets__ticket_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Ticket",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/v1/tickets/{ticket_id}/messages": {
      "post": {
        "description": "The reply lands in the conversation the ticket wraps, as `role='staff'`.",
        "operationId": "reply_to_ticket_api_v1_tickets__ticket_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketReply"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reply To Ticket",
        "tags": [
          "tickets"
        ]
      }
    },
...
    "/public/widget/tickets": {
      "post": {
        "description": "Open a ticket from the widget's contact form.\n\nGoes through the same public-key + origin allow-list + rate-limit chain as `/chat`, and\nthen a second, tighter gate of its own. The shared chain is not enough here: none of the\npublic key, the origin or the session id is anything a script cannot present a fresh one\nof, so on their own they shape ordinary traffic rather than stopping deliberate abuse.\nThis endpoint is the one worth abusing \u2014 it writes a row carrying an address the caller\nchose and up to 4000 characters that a human then reads \u2014 so it is also keyed on the\nclient address, which a caller cannot simply pick.",
        "operationId": "create_ticket_public_widget_tickets_post",
        "parameters": [
          {
            "description": "Public widget key (pk_...)",
            "in": "header",
            "name": "X-Chatbot-Key",
            "required": true,
            "schema": {
              "description": "Public widget key (pk_...)",
              "title": "X-Chatbot-Key",
              "type": "string"
            }
          },
          {
            "description": "Origin of the page hosting the widget, as attested by the browser",
            "in": "header",
            "name": "X-Widget-Site",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Origin of the page hosting the widget, as attested by the browser",
              "title": "X-Widget-Site"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ticket",
        "tags": [
          "widget"
        ]
      }
    }
  }
}